\name{bspec-package}
\alias{bspec-package}
\docType{package}
\title{
  Bayesian Spectral Inference
}
\description{
  This package functions to derive posterior distributions of spectral
  parameters of a time series.
}
\details{
  \tabular{ll}{
    Package: \tab bspec\cr
    Type:    \tab Package\cr
    Version: \tab 1.5\cr
    Date:    \tab 2015-04-23\cr
    License: \tab GPL (>=2)
  }
  The main functionality is provided by the \code{\link{bspec}()}
  function. .
}
\author{
Christian Roever <christian.roever@med.uni-goettingen.de>
}
\references{Roever, C., Meyer, R., Christensen, N.
  \href{http://dx.doi.org/10.1088/0264-9381/28/1/015010}{Modelling
    coloured residual noise in gravitational-wave signal processing}.
  \emph{Classical and Quantum Gravity}, 28(1):015010, 2011.
  See also \href{http://arxiv.org/abs/0804.3853}{arXiv preprint 0804.3853}.}
\keyword{ models }
\keyword{ package }

